Convert plugin to use SDK
