+++ .gitignore
@@ -3,3 +3,4 @@ node_modules
 # Local Netlify folder
 .netlify
 _next
+dist
