--- manifest.yml
@@ -1,23 +1,0 @@
-name: netlify-plugin-cloudinary
-inputs:
-  - name: cloudName
-    required: false
-    description: "Cloudinary Cloud Name - can be used as an input or environment variable via CLOUDINARY_CLOUD_NAME"
-  - name: deliveryType
-    required: false
-    description: "The method in which Cloudinary stores and delivers images (Ex: fetch, upload)"
-    default: "fetch"
-  - name: imagesPath
-    required: false
-    description: "Local path application serves image assets from"
-    default: "/images"
-  - name: folder
-    required: false
-    description: "Folder all media will be stored in. Defaults to Netlify site name"
-  - name: uploadPreset
-    required: false
-    description: "Defined set of asset upload defaults in Cloudinary"
-  - name: loadingStrategy
-    required: false
-    description: "The method in which in which images are loaded (Ex: lazy, eager)"
-    default: "lazy"
