+++ package.json
@@ -2,8 +2,11 @@
   "name": "netlify-plugin-cloudinary",
   "version": "1.2.0",
   "description": "Supercharge images on your Netlify site with Cloudinary!",
+  "type": "module",
   "main": "src/index.js",
   "scripts": {
+    "compile": "tsc --noEmit",
+    "build": "ntli build -b",
     "test": "jest"
   },
   "author": {
@@ -13,6 +16,7 @@
   },
   "license": "MIT",
   "dependencies": {
+    "@netlify/sdk": "^0.1.3",
     "cloudinary": "^1.27.1",
     "fs-extra": "^10.0.0",
     "glob": "^7.2.0",
@@ -22,8 +26,13 @@
   "devDependencies": {
     "@semantic-release/changelog": "^6.0.1",
     "@semantic-release/git": "^10.0.1",
+    "@types/fs-extra": "^11.0.1",
+    "@types/glob": "^8.1.0",
+    "@types/jsdom": "^21.1.1",
+    "@types/node-fetch": "^2.6.3",
     "dotenv": "^10.0.0",
     "jest": "^27.4.3",
-    "semantic-release": "^19.0.5"
+    "semantic-release": "^19.0.5",
+    "typescript": "^5.0.4"
   }
 }
