--- src/data/cloudinary.js
@@ -1,5 +1,0 @@
-const PREFIX = 'cld';
-module.exports.PREFIX = PREFIX;
-
-const PUBLIC_ASSET_PATH = `${PREFIX}-assets`;
-module.exports.PUBLIC_ASSET_PATH = PUBLIC_ASSET_PATH;
