+++ src/data/cloudinary.ts
@@ -1,0 +1,2 @@
+export const PREFIX = 'cld';
+export const PUBLIC_ASSET_PATH = `${PREFIX}-assets`;
