+++ src/data/errors.ts
@@ -1,0 +1,3 @@
+export const ERROR_CLOUD_NAME_REQUIRED = 'A Cloudinary Cloud Name is required. Please set cloudName input or use the environment variable CLOUDINARY_CLOUD_NAME'
+export const ERROR_NETLIFY_HOST_UNKNOWN = 'Cannot determine Netlify host, can not proceed with plugin.'
+export const ERROR_NETLIFY_HOST_CLI_SUPPORT = 'Note: The Netlify CLI does not currently support the ability to determine the host locally, try deploying on Netlify.'
